requirements.txt
matplotlib
numpy
pandas
scikit_learn
seaborn
folium
geopandas
openpyxl
lxml
geopy
mapclassify
adjustText
ipywidgets
openai
python-dotenv
streamlit
streamlit-antd-components
plotly
psycopg2-binary
sqlalchemy